Add an experimental program for cleaning up garbage owned by dwarfs.
As noted by Quietust on the forum, cleaning the owned flag from the
items does not actually remove the owner, but the item ceases to be
protected from manipulation by other dwarfs. This should be enough
to work around various bugs, like untouchable rotten food producing
miasma, or dwarfs littering the place with worn objects.