command clearvec: clear vector
Added command "clearvec" to plugin "vectors".  Takes a list of addresses
of vectors (or pointers to vectors), and sets each vector to zero length
by setting it's end to it's start.  Leaks memory, of course, but this
is just to poke at the game to see what happens when a vector's
contents disappear.