Split GitHub README from Sphinx index
The root of the documentation tree has a different function to the short
readme we want to display on GitHub, which now links to the main docs
(or Bay12 thread, or IRC).  It also renders properly on GitHub, and can
use live-updating status badges because we don't expect it to work
offline.