Clarify LICENSE.rst, deduplicate, add links
A simple list of components with the full text each time is not
particularly useful, or required.

The table lists components in alphabetical order (except DFhack, first),
with license and copyright.  Component names are hyperlinked to the
source project, which was not the case previously.

Licenses which are required to be included are, but once only - an
improvement over four zlib, 3 BSD and five (!) MIT blocks.  This and the
table satisfies the requirements while remaining readable.