Add Introduction.rst, list builtins in Core.rst
The Introduction is the first page to read, and thus explains what
DFHack is, installation, basic use, and troubleshooting.

This frees up Core.rst for a detailed list of builtin commands (meant to
be complete, but I might have missed some), some other key commands, and
a discussion of init files.  I also put the random notes that don't fit
anywhere at the bottom of this file.