Document menu-mouse plugin
Closes #500, by merging Felton's plugin.

menu-mouse is not perfect, and can have some strange interactions with
other GUI-altering tools.  However it does work and provides a very
valuable function for some players who would be willing to live with the
downsides.

Given that Felton seems to have gone offline for good, I think it should
be in the DFhack repo where further development might take place.