List valid options for caste
The wiki provides no indication as to how to even determine the values for <caste> and after a few hours of looking, I couldn't find a good way either, other than just guessing from stuff in the raws.  Since we're iterating the entire list of valid castes anyway, we might as well collect them to a string we can output to help the user.