track remaining job_item quantity
instead of wiping the job_item filters during finalization. this allows
buildings partially completed with buildingplan, but finished with
buildingplan disabled, to complete semi-normally. they'll be built with
the correct number of each item, but the items might not be in the
correct order, and the 'rough' attribute might not be correct. However,
I think this is the best we can do in that (uncommon) situation.

This change also enables a technique I have in mind for the future for a
"prefer blocks" setting, where multiple item attachment tasks are
registered with different item vectors, and the extra tasks are popped
as invalid when the job_item quantity is detected as being zero.