index by id instead of bld pointer
because pointers are not guaranteed to be unique. if you allocate a
building, destroy it, and then allocate another building, the two
buildings can have the same pointer, but they'll have different ids.

this solves the problem that a new building could get erroneously
reported as already being registered with buildingplan.