Remove extra paragraph from generated changelogs, restore paragraph margins
The CSS (changed in bca76b8f) was removing space between actual paragraphs in
lists. This was intended to address excess padding in changelogs, but that is
resolved here by removing blank lines surrounding nested lists. This still
displays properly on GitHub/Reddit and presumably other Markdown implementations
as well.