Stop reversing changelog entries
This makes entries show up in the same order in the generated docs as in changelog.txt

The only other significant difference this causes is that when notes from multiple prereleases are combined into one stable release, changes from the newer version will show up first now (i.e. in the same order as reading the changelog from the top down), but this has minimal impact.