Disable shallow clones in .gitmodules
Shallow clones cause issues with `git push --recurse-submodules=check|on-demand`
such as in https://github.com/DFHack/dfhack/runs/3307192897?check_suite_focus=true:

```
The following submodule paths contain changes that can
not be found on any remote:
  depends/luacov

Please try

    git push --recurse-submodules=on-demand

or cd to the path and use

    git push

to push them to a remote.
```

(in this case, the changes *are* on the remote, and on-demand doesn't help)

Our submodules also aren't currently very large on disk even with full clones -
libexpat is the largest:

616K    .git/modules/depends/clsocket
744K    .git/modules/depends/luacov
924K    .git/modules/depends/xlsxio
1.8M    .git/modules/depends/libzip
3.7M    .git/modules/depends/jsoncpp
8.2M    .git/modules/depends/libexpat

If a shallow clone of submodules is desired, `git clone --shallow-submodules` is
still an option.