hive-color.h
Hives are normally shown as yellow, regardless of tool material. As with nestboxes, this should change how the games show hives, making it possible to see what material the hives are made of.

Can't really confirm if it works or not yet, but should work since there is little difference between how buildings derived from tools work, especially between nest-boxes and hives.