[manipulator] Use more muted colors to help cursor stand out
This is similar to the original color scheme, but uses a brighter cursor
color (YELLOW instead of BLUE) and darker colors for the rest (BLUE
instad of GREY, BROWN instead of CYAN, GREY instead of WHITE). The
result is that the cursor is easier to see, while everything else is
hopefully just as legible as before.