[tweak/stable-cursor] Fix for when default menu is off
When the default menu is off, the viewport shifts by 16 (half the menu
width), but getViewportPos() does not report that until after the
interpose vmethod is over. So the view is always off by 16 in the x
direction.

The check_menu() function is a simplified version of what is in
Gui::getDwarfmodeViewDims().