[tweak] Fix stable-cursor when no menu at right map edge
When the view is near the right edge of the map, and the default menu is
off, the view never stays exactly at the right edge when the menu
disappears. It is always shifted over half the menu's width from the
right edge.

Finding the right value involves duplicating some functionality from
Gui::getDwarfmodeViewDims(), because that hook is cached and does not
update fully during this INTERPOSE hook. In particular, getting the
actual viewport width (accounting for area map & menu).