[tweak] stable-cursor fix for Build sidebar when menu off
The Build sidebar has an extra step before the cursor shows up
(selecting the type of building before it can be placed). This requires
an extra check, since the menu is already forced onscreen when the
cursor appears.

I've also simplified the if/else conditions to remove redundancy.
There's no need to clear the last cursor position if it's already
invalid. Now it is cleared once, immediately after it's used.