Modernizes string comparisons by a degree
also fixes a possible bug, on L#1086-'88
See how the variable `you` isn't actually used no clue how it survived all this time.

I've gone ahead and presumed the intention was to not repeat the
comparison performed on line 1077, so I've used the string literal
"You " in the comparison instead of that.