Remove library/military silver crossbow work-order
It is not possible to create a work-order for making silver crossbows,
so the silver crossbow recipe belongs in the alternative military library
with other artifact/non-craftable weapon recipes.

It's already present in the library/military_include_artifact_materials
work-orders, so simply needed removal from the vanilla library/military
orders version.