Add "faststart" plugin to make DF start faster
In particular, it makes the game's "Loading..." screen animate
as quickly as possible, shortening it from around 10 seconds to
slightly more than 1 second.

A conditional build setting makes it skip the animation as well,
making it slightly faster yet.

Ideally, this should become part of the Tweak plugin, but we're
not building that right now.