Make Items::createItem handle multiple output
This changes the function signature to return a list of created items
via the first parameter and a boolean "success" code. It also moves
the "unit" parameter to the beginning and adds "growth_print" to the
end (which is required for PLANT_GROWTH items).

Also adjusts 'createitem' plugin to use Items::createItem().

Should fix #2063