gui.dflayout: support for "growing" UI elements
The UI elements that were initially supported (toolbars and their
buttons) were all fixed-sizes and had inset values that were
non-decreasing as the interface width grew.

The new functionality supports UI elements that grow with the size of
the interface area and have inset values that may increase or decrease
as the interface size (or other DF state) changes.

The "item list" part of two tabs from the DF info window are given as a
demonstration "UI element": the Orders tab and the Places/Zones sub-tab.