gui.dflayout: export getUIElementFrame and getRelativePlacement
The first is the previously-internal wrapper around calling `.frame_fn`.
This should help with callers treating the UI element values as opaque,
since `frame_fn` was the most interesting bit there.

The second is the previously-internal "relative positioning" function
used in `getOverlayPlacementInfo`. It should be useful in other contexts
that would like to do element-relative position calculations, but
without the extras stuff that overlay-positioning requires.