info sort overlay: use interface width in resize_overlay
The DF info window is based on the interface area, not the full window
area. So the sort info overlays should also work with the interface
width instead of the window width.

If DF is configured to use an interface percentage smaller than 100,
then the interface area will be narrower than the full window once the
window is resized beyond the minimum width. Basing the info sort overlay
size on the window size can let some UI elements extend outside the
footprint of the info window.

For example, when the DF interface percentage is set to 90, the search
box for the Task tab
- is too close to the info window border at window widths 115 and 147
- shares a edge with the info window border at window widths 116 and 146
- extends beyond the info window border for window widths 117-145